--- a/gulpfile.js
+++ b/gulpfile.js
@@ -49,6 +49,11 @@ gulp.task('watch:core', core.watch)
 gulp.task('watch:studio', gulp.series([ui.cleanStudioAssets, ui.createStudioSymlink, ui.watchStudio]))
 gulp.task('watch:admin', ui.watchAdmin)
 
+gulp.task(
+  'watch:ui',
+  gulp.parallel([ui.watchAdmin, gulp.series([ui.cleanStudioAssets, ui.createStudioSymlink, ui.watchStudio])])
+)
+
 gulp.task('clean:node', cb => rimraf('**/node_modules/**', cb))
 gulp.task('clean:out', cb => rimraf('out', cb))
 gulp.task('clean:data', cb => rimraf('out/bp/data', cb))