--- a/src/bp/ui-admin/package.json
+++ b/src/bp/ui-admin/package.json
@@ -58,6 +58,7 @@
     "@types/react-select": "^2.0.19",
     "@types/reactstrap": "^8.0.1",
     "cross-env": "^5.2.0",
+    "hard-source-webpack-plugin": "^0.13.1",
     "node-sass-chokidar": "^1.3.4",
     "npm-run-all": "^4.1.5",
     "react-app-rewired": "^2.1.0"