--- a/src/bp/ui-studio/package.json
+++ b/src/bp/ui-studio/package.json
@@ -87,6 +87,7 @@
     "expose-loader": "^0.7.1",
     "file-loader": "^1.1.11",
     "find-with-regex": "^1.1.3",
+    "hard-source-webpack-plugin": "^0.13.1",
     "html-webpack-plugin": "^3.2.0",
     "immutable": "^4.0.0-rc.12",
     "node-sass": "^4.8.3",