I'm trying something to speed up the build process of admin / studio. The plugin adds a layer of cache for node_modules and I got a nice improvement on build times (first build is still long, but subsequent ones are a lot quicker):

Admin:  build time went from 17s to 14s
Studio: Build time from 16s to 5s

Working on the studio is a lot quicker. Also added a timestamp when watching studio, so you can see when your last change has been compiled

And, use `yarn cmd watch:ui` to launch both admin and studio in watch mode.

Can you guys try it out to see if you have any issue ?